+++ lib/loaders/KCircularLoader.vue
@@ -11,6 +11,7 @@
 
   <transition name="ui-progress-circular--transition-fade">
     <div
+      v-if="showLoader"
       class="ui-progress-circular"
       :class="[classes, { delay }]"
       :style="{ 'width': size + 'px', 'height': size + 'px' }"
@@ -92,6 +93,13 @@
         default: 'indeterminate', // 'indeterminate' or 'determinate'
       },
       /**
+       * Whether the loader should be displayed or not. It needs to be used instead of `v-if/v-show` for `minVisibleTime` to work.
+       */
+      show: {
+        type: Boolean,
+        default: true,
+      },
+      /**
        * For determinate loaders, value between 0 and 100 representing percentage completion
        */
       progress: {
@@ -104,6 +112,13 @@
       delay: {
         type: Boolean,
         default: false,
+      },
+      /**
+       * Do not hide the loader until `minVisibleTime` in milliseconds. Useful to avoid jarring UX when the actions finishes very fast. In comparison to `delay`, `minVisibleTime` emphasizes that an action associated with the loader is indeed taking place. `show` needs to be used instead of `v-if/v-show` for this to work.
+       */
+      minVisibleTime: {
+        type: Number,
+        default: 0,
       },
       /**
        * Overall size of the loader in pixels
@@ -121,7 +136,18 @@
       },
     },
 
+    data() {
+      return {
+        isFrozen: false,
+        freezeTimeoutId: null,
+      };
+    },
+
     computed: {
+      showLoader() {
+        return this.show || this.isFrozen;
+      },
+
       classes() {
         return [`ui-progress-circular--type-${this.type}`];
       },
@@ -158,6 +184,28 @@
       },
     },
 
+    watch: {
+      show(oldVal, newVal) {
+        if (oldVal === newVal) {
+          return;
+        }
+        this.freezeLoader();
+      },
+      freeze(oldVal, newVal) {
+        if (oldVal === newVal) {
+          return;
+        }
+        if (newVal === 0 || newVal < 0) {
+          clearTimeout(this.freezeTimeoutId);
+        } else {
+          this.freezeLoader();
+        }
+      },
+    },
+    mounted() {
+      this.freezeLoader();
+    },
+
     methods: {
       moderateProgress(progress) {
         if (isNaN(progress) || progress < 0) {
@@ -169,6 +217,15 @@
         }
 
         return progress;
+      },
+
+      freezeLoader() {
+        if (this.show && this.minVisibleTime > 0) {
+          this.isFrozen = true;
+          this.freezeTimeoutId = setTimeout(() => {
+            this.isFrozen = false;
+          }, this.minVisibleTime);
+        }
       },
     },
   };